requirements.txt
black==25.1.0
colorama==0.4.6
flake8==7.2.0
isort==6.0.1
mypy==1.15.0
piexif==1.1.3
pillow==11.2.1
pylint==3.3.7
shellingham==1.5.4
typer==0.15.4
